Simplified the auth code path - but also unified it between the two secure paths
