+++ application/backend/src/api/auth/session-cookie-constants.ts
