+++ application/backend/src/api/routes/trpc-context.ts
@@ -5,6 +5,15 @@ import { UsersService } from "../../business/services/users-service";
 import { ReleaseService } from "../../business/services/release-service";
 import { getServices } from "../../di-helpers";
 
+/**
+ * Create the base context for our TRPC calls, provided useful base
+ * services and the underlying request and response for the call.
+ *
+ * Note: other route middleware may add values to this context such
+ * as the authenticated user.
+ *
+ * @param opts
+ */
 export const createContext = async (opts: CreateFastifyContextOptions) => {
   const { edgeDbClient, settings, logger } = getServices(container);
   return {
